+++ client/src/main/java/net/oneandone/stool/client/cli/Attach.java
@@ -15,6 +15,7 @@
  */
 package net.oneandone.stool.client.cli;
 
+import net.oneandone.inline.ArgumentException;
 import net.oneandone.stool.client.App;
 import net.oneandone.stool.client.Globals;
 import net.oneandone.stool.client.Project;
@@ -26,31 +27,64 @@ import java.util.List;
 
 public class Attach extends ProjectCommand {
     private final String stage;
+    private final String pathOpt;
 
     public Attach(Globals globals, FileNode project, String stage) {
         super(globals, project);
-        this.stage = stage;
+
+        int idx;
+
+        idx = stage.indexOf("=");
+        if (idx == -1) {
+            this.stage = stage;
+            this.pathOpt = null;
+        } else {
+            this.stage = stage.substring(0, idx);
+            this.pathOpt = stage.substring(idx + 1);
+        }
     }
 
     @Override
     public void doRun(FileNode project) throws Exception {
         Project backstage;
+        List<FileNode> wars;
+        String nameAndServer;
 
-        List<Reference> found;
-
-        found = globals.servers().list(stage);
-        switch (found.size()) {
-            case 0:
-                throw new IOException("no such stage: " + stage);
-            case 1:
-                break;
-            default:
-                throw new IOException("stage ambiguous: " + stage);
-        }
         backstage = Project.lookup(project);
         if (backstage == null) {
             backstage = Project.create(project);
         }
-        backstage.setAttached(new App("TODO", found.get(0)));
+        if (pathOpt == null) {
+            wars = backstage.wars();
+            if (wars.isEmpty()) {
+                throw new ArgumentException("no wars found - did you build your project?");
+            }
+            if (wars.size() != 1) {
+                throw new IllegalStateException("TODO: too many wars");
+            }
+            for (FileNode war : wars) {
+                nameAndServer = Create.app(war) + "." + stage;
+                checkStage(nameAndServer);
+                backstage.setAttached(new App(checkStage(nameAndServer), war.getRelative(project)));
+            }
+        } else {
+            project.findOne(pathOpt);
+            backstage.setAttached(new App(checkStage(stage), pathOpt));
+        }
+    }
+
+    private Reference checkStage(String nameAndServer) throws IOException {
+        List<Reference> found;
+
+        found = globals.servers().list(nameAndServer);
+        switch (found.size()) {
+            case 0:
+                throw new IOException("no such stage: " + nameAndServer);
+            case 1:
+                return found.get(0);
+            default:
+                throw new IOException("stage ambiguous: " + stage);
+        }
+
     }
 }